Rebase xmrig-nvidia-2.6.1 version

+++ src/nvidia/cuda_core.cu
@@ -495,7 +495,7 @@ void cryptonight_core_gpu_hash(nvid_ctx* ctx, uint32_t nonce)
 }
 
 
-void cryptonight_gpu_hash(nvid_ctx *ctx, xmrig::Algo algo, int variant, uint32_t startNonce)
+void cryptonight_gpu_hash(nvid_ctx *ctx, xmrig::Algo algo, int variant, uint64_t startNonce)
 {
     using namespace xmrig;
